pyproject.toml
[project]
name = "rebalancer"
version = "0.1.0"
description = "Automated portfolio rebalancer with backtesting and reporting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "plotly>=5.0",
    "click>=8.1",
]

[project.scripts]
rebalancer-simulate = "rebalancer.cli:simulate"
rebalancer-daily = "rebalancer.cli:daily_check"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rebalancer"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
